remove emotion from Folder.tsx
